alacritty: new window script
using open to start new windows on a mac starts up a new process for each window: spammy.

With this, I configure mac built-in Shortcut program with keybinding shift-cmd-N, and then it triggers this script.  The script gets the current running process and attempts to notify it to create a new window